test updated byteplay with support for 3.6
