+++ rootpy/extern/byteplay3/README.rst
@@ -1,0 +1,1 @@
+https://github.com/serprex/byteplay
