+++ rootpy/utils/inject_closure.py
@@ -1,10 +1,11 @@
 import types
-
 import sys
+
 if sys.version_info[0] >= 3:
     from ..extern import byteplay3 as byteplay
 else:
     from ..extern import byteplay
+
 from ..extern.six.moves import range
